tests/fbcon_fbt: Add and use psr_long_wait_exit() to test updates in PSR
After getting back to fbcon it tests if PSR had a update but since
fbcon can take a while it have some big sleep in place, so what it
is actualy testing is if fbcon is updating the screen.
In this test the update of the screen comes from the fbcon cursor
blinking, the problem is that 40ms is a small interval to detect
cursor blinking and the test can fail some times, so a bigger timeout
to wait for exit is need.

Cc: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>