test/chamelium: Add HDMI slow unplug workaround
HDMI specification states that the HPD should be shorter then the
other pins, so slowing unpluging will cause a HPD interruption that
can be processed before the other pins lose contact, so kernel will
be able to read the data lines and will keep the connector as
connected.
A workaround was added in kernel to retry the hotplug detection in
case we detect a hotplug and there is not change in the connector
state, so lets add a test to simulate that issue using chamelium.

Cc: Imre Deak <imre.deak@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>