tests/kms_frontbuffer_tracking: Improve tiling test coverage
Lets add some more tiling tests, if tiling is supported by FBC draw
some rectangles and compare the CRC against benchmark if not
supported run the test to guarantee that FBC is disabled.

This is a preparation for when kernel will allow FBC to be enabled
without fences, so we can better test linear and Y tiled
frontbuffers.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>