Revert "tests/kms_cursor_legacy: Handle cursor movements as slow patch for display 9+"
Cursor movements will not be handled as fastsets right now because
we have performance issues as X11 based compositors do hundreds of
cursor movements per frame and if we sync all atomic commits it will
case a huge lag.
Kernel is already handling it correctly but we missed revert the
IGT side.

This reverts commit 984be097711f99fa7e9b3681da2adbd2ccfd6f55.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>