build: check that outb is present in io.h
With glibc if io.h exists, inb/outb are always defined.
However on musl, io.h always exists but it may not define inb/outb.

Thus, igt-gpu-tools builds with musl on non-x86 platforms will fail to
link. Fix this by checking for both io.h and that outb() is defined.

Signed-off-by: Bernd Kuhls <bernd@kuhls.net>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>