tests/kms_async_flip: Speed up the CRC test on Intel hw
Go back to using igt_draw mmap stuff to do the mid-test
rendering in the CRC subtest to make it faster. Going
via all the cairo machinery is very very slow, but we'll
leave that codepath in place for non-Intel hw. This makes
the test pass on a DG2 here as we can now produce enough
frames to have some confidence in the results.

Note that some extra care is apparently needed with the order
in which things are done. If we pin the buffers to the display
first the mmap is likely to fail (due to the FB getting pinned
outside the reach of the small BAR presumably). Curiosuly the
mmap itself succeeds but we get a SIGBUS when trying to acccess
the buffer contents. If we touch the FBs first via mmap and only
then pin them to the display everything works correctly.

v2: Use igt_draw_supports_method() so it works on xe (Bhanu)

Cc: Melissa Wen <mwen@igalia.com>
Cc: Alex Hung <alex.hung@amd.com>
Cc: André Almeida <andrealmeid@igalia.com>
Cc: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Fixes: 8ba806e7e196 ("tests/kms_async_flips: Support more vendors")
Reviewed-by: Karthik B S <karthik.b.s@intel.com>
Tested-by: Karthik B S <karthik.b.s@intel.com>
Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/8561
Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/8247
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>