tests/xe: Remove xe_perf_pmu
Group busyness uapi implemented in PMU in the kernel has been discontinued
and will be replaced by a different group busyness uapi and a different
implementation. Because of this, we are proposing to remove the PMU code
from the kernel, till the correct uapi is implemented. Therefore drop the
corresponding IGT too, since it caters to the old uapi.

Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Acked-by: Aravind Iddamsetty <aravind.iddamsetty@linux.intel.com>