tests/kms_vrr: Fix pipe/output combo validity constraint
Before starting the subtest, we are checking the selected pipe/output
combo validity with default mode, but inside the subtest we are using
the mode with highest resolution.

Few panels exposes 4K as a default mode even though it supports 8K,
in this scenario test may misbehave.

So, check the pipe/output combo validity with the mode that we are
actually going to use in the subtest.

V2: Reword the commit message

Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Reviewed-by: Mitul Golani <mitulkumar.ajitkumar.golani@intel.com>