tests/kms_vrr: Tweak flipline subtest
As per the spec: If there is no flip request with in Vmax,
Hardware will terminate the Vblank at Vmax.

If we flip with the interval > Vmax, the expected refresh rate
could be: (Vmax + time left for the next flip + Vmin), and it
is very difficult to generalize. For low refresh rate panels
(like 40 - 60 Hz), result is not falling into the threshold.

Hence, tweak the test to increase the flip request interval.

Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Tested-by: Vidya Srinivas <vidya.srinivas@intel.com>
Reviewed-by: Mitul Golani <mitulkumar.ajitkumar.golani@intel.com>