tests/intel/gem_exec_schedule: Update for MTL RCS/CCS w/a
The RCS/CCS w/a means that if spinners are sent to both RCS and CCS
engines then the system is dead and requires a reset. Even
pre-emptible spinners cannot be pre-empted due to the nature of the
w/a. So, do not get into that situation.

v2: Expand code comment (review feedback by Ashutosh)
v3: Fix style issue and broken email address (review feedback from
Kamil)

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>