tests/kms_lease: Cache xe_device info for lease fd
In case of xe device for lessor fd, the xe_device struct gets
created and cached during drm_open. But for the lessee fd,
the xe_device info struct is not created and cached.

This causes problem in prepare_crtc() for lessee with xe, which
needs the xe_device info struct. So, create and cache the xe_device
struct info for lease fd, when we make lease and remove it from
cache while terminating the lease.

v2: -Update description and move logic to make_lease call. (Ankit)
    -Add xe_device_put, while terminating the lease. (Ankit)

Signed-off-by: Mohammed Thasleem <mohammed.thasleem@intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>