lib/igt_kms: Correct the check for igt_require
The function 'igt_debugfs_dir()' returns -1, if it fails and a non-negative
fd otherwise. However in the function igt_get_max_dotclock() the condition
to check the debugfs directory uses igt_require (dir), which fails even if
the dir is 0.
So correct the igt_require check for a valid igt_debugfs fd.

v2: Correct the logic in igt_require.

Signed-off-by: Mohammed Thasleem <mohammed.thasleem@intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>