tests/intel/xe_exec_store: account for prefetch on Xe2
Due to batch pre-fetch being quite aggressive on Xe2, we need
an extra 4K page at the end of the batch to ensure HW doesn't
pre-fetch beyond the boundaries of the batch itself, hitting
an invalid ppGTT page, resulting in CAT errors. Fix this by
accounting for the xe_cs_prefetch_size() when calculating
the size of the batch.

Suggested-by: Matthew Auld <matthew.auld@intel.com>
Signed-off-by: Jagmeet Randhawa <jagmeet.randhawa@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>