tools/intel_gpu_top: Fix near full percentage bar formatting
Fix a bug where 1) the end vertical separator element would not be printed
if the progress bar portion was all filled by the progress bar characters
(no trailing spaces), and 2) the numerical overlay would be skipped to.

The bug would also shift the layout of following UI elements since the
progress bar would not be consuming all the allocated horizontal space.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reported-by: anonymoustranquillity@proton.me
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>