lib/intel_blt: use BYTE_COPY mode on xe2
Currently we see various CAT errors and timeouts on xe2 when running
xe_copy_basic. If selecting PAGE_COPY mode for MEM_COPY, the page size
is defined as 256B block on xe2. Also the width should be defined in
number of pages, and not bytes, which likely explains the CAT errors.
However the caller in xe_copy_basic is not using a size that is aligned
to 256B, so rather just select the more general BYTE_COPY mode instead,
which should work for any reasonable number of bytes. Also bit 19
doesn't look to be supported on PVC, where byte copy is all we seem to
have. In addition the matrix copy also expects byte copy.

BSpec: 57561
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Acked-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>