scripts/igt_doc.py: allow using multiple config files at the same time
When generating testlists meant to test multiple testplan configs at
the same time, we need to group tests from different testplans.

Add support for it by letting the command be called with multiple
config files.

It should be noticed that, when --json or --rest options are used,
igt_doc.py can't support multiple config files, as both options
generate just one output file which is based on what it is
defined on a single config file.

Printing ReST output at the screen works, but the results will
be independently generated, so they'll be grouped per input file.
As this may not be what the user expects, add a warning at the end.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>