lib/kunit: Enter legacy path only if applying filters fails
If loading the base KUnit module in list mode succeeds but our KTAP
parser or test suite filtering returns an empty list of test cases then,
instead of calling igt_skip, we now unintentionally fall back to legacy
mode as if the module didn't support filter parameters.

Fall back to legacy path only when kunit_get_tests() fails to load the
base KUnit module with non-default filters, and not when it returns an
empty list of test cases.

As an additional benefit from this approach, we now have a single TODO
place in the code to be fixed when the fallback is no longer needed.

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>