lib/kunit: Fix API documentation
The only IGT KUnit global function igt_kunit() accepts a "name" argument
documented as "the name of subtest, if different from that derived from
module name".  But in practice, we use that argument, if not NULL, as a
name of test suite to be executed.  We pass that value to kunit_get_tests()
as a filter that selects only test cases from that test suite.  We also
use that string, if not NULL, as subtest name, which we otherwise just
derive from module name.

Fix the documentation, and also use better names for the argument and a
local variable initialized as its copy.  For clarity, propagate those
new names as argument names down to lower level functions.

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>