tests/intel/kms_frontbuffer_tracking: Invalidate cached stuff
The test pipe-fbc-rte updates the primary mode parameters for each valid
dynamic test case. Because of that, we might endup with invalid cached
data due to differences between the initial state of prim_mode_params
defined from the beginning of the test program and the possibly changed
state after pipe-fbc-rte.

As an example, in a specific environment, the command

  ./build/tests/kms_frontbuffer_tracking \
      --run-subtest pipe-fbc-rte,fbc-1p-primscrn-pri-indfb-draw-mmap-wc

would result in fbc-1p-primscrn-pri-indfb-draw-mmap-wc failing because
it would try to read CRC from pipe B while the test was being actually
done in pipe A.

Another potential issue worth noting is that even pipe-fbc-rte could
similarly fail if the set of dynamic subtests spanned across multiple
pipes.

Let's fix that by making sure that cached stuff that would depend on the
primary mode parameters gets properly invalidated when prim_mode_params
is the target of init_mode_params(). This should fix the issues
mentioned above and also future-proof the code for any future test that
would need to modify the prim_mode_params.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>
Reviewed-by: Ravi Kumar Vodapalli <ravi.kumar.vodapalli@intel.com>
Link: https://lore.kernel.org/r/20240209135526.135380-2-gustavo.sousa@intel.com
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>