tests/intel/gem_watchdog: Reduced timeouts for worst case scenario
The watchdog test reduces the watchdog timer from 20s to 1s and then
uses a 5s timeout waiting for the watchdog to do its stuff. This works
fine in general, but if an engine reset is required by a context that
is actually dead for real then a pre-emption timeout must be factored
in. For RCS/CCS engines, that timeout is 7.5 seconds by default. Thus,
the test timeout expires first and the test fails.

Normally, the system is not so dead when running this test as to
require an engine reset. A simple pre-emption works fine for the
spinner contexts that is uses. However, there is a hardware workaround
coming which prevents context switches when both RCS and CCS are busy.

So add an explicit override of the pre-emption timeout as well as the
watchdog timeout. That will allow the test to keep working after the
new w/a lands.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>