tests/prime_nv_test: switch over to intel_buf_init()
We want to get rid of intel_buf_init_using_handle(), in favour of always
passing in the real bo size, otherwise it is quite tricky to figure out
what exactly to set for buf->bo_size (which eventually get plugged into
vm_bind) when the caller is the one who created the bo. Trying to guess
the page alignment is unsafe without knowing more about the actual
object.  In this case it seems simplest to switch over to
intel_buf_init. Should be no functional change here.

Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Acked-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>