lib/intel_bufops: repurpose intel_buf_create_using_handle_and_size()
From an api pov user this should be the same as intel_buf_create, except
with existing buffer handle and bo_size. There are no current users for
this, but in upcoming patch we will remove
intel_buf_create_using_handle() and convert all users over to this api,
such that we don't have to guess the bo_size deep in the lib code.

Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>