scripts/igt_doc.py: fix intelci testlist generation for complex cases
The most complex case is when a test is included or block listed, via
GPU and "Excluded GPU platform" fields, as the same test might be
on both lists.

Currently, the handling logic doesn't go though all GPUs, which ends
supressing tests from testlists if the GPU is not mentioned.

Fix it, while simplifying the logic.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>