lib/rendercopy: Fix Xe2 pixel shader
On Xe2 start and coefficient for barycentric positions are passed
in different location of grf register (r6). I wasn't able to find
this information explicitly in the documentation so I've dumped
all registers involved in the operation and deduce r6 values
meaning. This means I may be wrong but tests which are previously
failing on render-copy now are working fine.

Cc: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Cc: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Link: https://lore.kernel.org/r/20240226114511.51753-1-zbigniew.kempczynski@intel.com
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>