tests/intel/xe_gt_freq: re-introduce exec tests of xe_gt_freq
The tests freq_range_exec, freq_fixed_exec were removed due
to a design flaw which caused failure in tests. The workload execution
finished earlier before the frequency validation when gt was set to
run at a higher frequency.

These tests are being reintroduced with a spinner based approach
in this patch to avoid such failures.

Signed-off-by: Riana Tauro <riana.tauro@intel.com>
Reviewed-by: Sujaritha Sundaresan <sujaritha.sundaresan@intel.com>
Signed-off-by: Anshuman Gupta <anshuman.gupta@intel.com>