tests/intel/gem_spin_batch: Skip compute engines on MTL due to w/a
There is a w/a on MTL that prevents concurrent execution of multiple
processes on RCS and CCS. The 'spin-all-new' test is explicitly
testing such execution across all engines. So either the entire test
must be skipped on MTL or just the compute engine must be skipped. Do
the latter as that at least leaves some test coverage in place.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>