tools/intel_vbt_decode: Dump MIPI sequences in the order they appear in VBT
Dump the MIPI sequence blocks in the order they appear
in the VBT (as opposed to being ordered by the sequence type).

This makes it easier to compare the hexdump vs. the decoded
output, which is what I tend to do occasionally when trying to
check whether we are decoding things correctly or not.

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>