tests/intel/xe_evict: Skip test if smem less than lmem
Currently the tests are failing with OOO memory error.
Test requires SRAM to be as big as VRAM. For example, small-cm creates
(448 / 2) BOs with a size (1 / 128) of the total VRAM size. For
simplicity ensure the SRAM size >= VRAM before running this test.
Added code to skip the test if the system memory size is less than local
memory size.

v2: Added memory size in the skip print.(kamil)
Also added reason as a comment for the skip.(Matt)

Cc: Matthew Brost <matthew.brost@intel.com>
Cc: Venkata Ramana Nayana <venkata.ramana.nayana@intel.com>
Signed-off-by: Priyanka Dandamudi <priyanka.dandamudi@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Sai Gowtham Ch <sai.gowtham.ch@intel.com>