lib/igt_kmod: stop using KMOD_REMOVE_FORCE
The flag to force remove a module should never be used. The flag is
intended for end users to decide if they want to have a broken system
left after that. Maybe it's a "controlled breakage". It's not something
for a testsuite to use.

If the kernel ever resorts to honoring the flag (due to module's kref
not being 0), we will have a tainted kernel, potential leaks and memory
corruption. It's better to just handle the error returned by libkmod and
fix the root cause.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>