lib/xe_spin: Fix MI_STORE_REGISTER_MEM instruction usage
The MI_STORE_REGISTER_MEM macro in IGT's instruction header is the "old"
version of this define, intended for use on pre-gen8 platforms since it
bakes in the fixed 3-dword instruction length in the instruction header.
For gen8 and beyond (i.e., every platform the Xe driver supports), the
MI_STORE_REGISTER_MEM_GEN8 macro should be used instead, which properly
bakes in the 4-dword instruction length.  Since the xe_spin library was
using the old macro, but then manually |'ing in a "2" at the end of the
expression, it was effectively winding up with a field value of 3 (which
would represent a 5-dword instruction).

Switch xe_spin to using the "GEN8" macro and drop the redundant "| 2"
which is already handled by the macro itself.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Nirmoy Das <nirmoy.das@intel.com>