lib/kunit: Time out promptly on missing KTAP report
If a test provides a subtest name that doesn't match any test suites
provided by the requested KUnit test module then no KTAP report appears in
dmesg, not even an empty one as one may expect.  As a consequence, we now
loop endlessly around reading potential lines of the missing report from
/dev/kmsg, until killed by IGT runner on timeout.

When trying to collect names of test cases from a KTAP report generated in
all skip mode, set an alarm that fires up 10 seconds after we start
waiting for the report, interrupts blocking read() if pending, and
terminates the subtest with SKIP result.

As soon as we have collected a non-empty list of test cases, we may as
well expect a non-empty KTAP report from actual execution of those test
cases, assuming successful load of the KUnit test module in execution
mode.  Then, there is no need to set up a similar timeout before we start
to extract and parse that report.

Suggested-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Acked-by: Jonathan Cavitt <jonathan.cavitt@intel.com>