tests/xe_exec_threads: Fill in GT field for second balancer thread
The balancer subtests spawn two pthreads per engine class if there are
multiple instances of the class.  The GT field of the data structure is
filled in properly for the first thread, but not for the second,
effectively leaving it set to "0."

For platforms with standalone media, this will result in failures when
the thread tries to find the instances of a media class on GT0 and trips
the "igt_assert(num_placements > 1)" assertion in test_balancer().

Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Link: https://lore.kernel.org/r/20240320192356.359327-1-matthew.d.roper@intel.com
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>